Backend/base/requirements.txt
django==4.2
djangorestframework==3.14.0
mysqlclient==2.2.6
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
Pillow==11.1.0